ocr model will load from another directory
