+++ src/_models/__init__.py
@@ -1,0 +1,6 @@
+from ._load_ocr import _get_ocr_model, _release_ocr_model
+
+__all__ = [
+    "_get_ocr_model",
+    "_release_ocr_model",
+]
